Modif sur la methode getLocation(), l'app ne crash plus si location == null

+++ GSB_Mobile/app/src/main/java/com/example/leo/gsb_mobile/ui/MyViewHolder.java
@@ -45,7 +45,7 @@ public class MyViewHolder extends RecyclerView.ViewHolder{
         textViewNomM.setText(object.getNomMedecin());
         textViewPrenomM.setText(object.getPrenomMedecin());
         textViewAdresseC.setText(object.getAdresseCabinet());
-        textViewDistance.setText((int) object.getDistance());
+        textViewDistance.setText(object.getDistance());
         idMedecin = object.getIdMedecin();
     }
 }
